# openapi.yaml
openapi: 3.0.3
info:
  title: Final API
  version: 1.0.0
  description: Course task tracker
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /reports:
    post:
      summary: Create a new report
      responses:
        "202":
          description: Report job has been sent to queue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
              example:
                jobId: report-17
                status: pending
                statusUrl: /reports/report-17
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /reports/{id}:
    get:
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The report ID from the URL path.
          schema:
            type: string
            example: report-17
      responses:
        "200":
          description: Single report response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
              example:
                id: report-17
                studentId: djs001
                status: completed
                downloadUrl: /downloads/report-djs001-report-17.json
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          description: The report with the provided id was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Not found
components:
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
          example: report-17
        studentId:
          type: string
          example: dhs001
        status:
          type: string
          example: completed
        downloadUrl:
          type: string
          nullable: true
          example: /downloads/report-djs001-report-17.json
      required:
        - id
        - studentId
        - status
    ReportResponse:
      type: object
      properties:
        jobId:
          type: string
          example: report-17
        status: 
          type: string
          example: pending
        statusUrl:
          type: string
          example: /reports/report-17
      required:
        - jobId
        - status
        - statusUrl
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT bearer token in the Authorization header using format `Bearer <token>`.
  responses:
    UnauthorizedError:
      description: Unauthorized response for missing, expired, malformed, or invalid tokens.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "The token is missing, invalid, expired, or malformed."
...